requirements.txt
# Minimum supported Python: 3.8+
cocotb==1.8.2
cocotb-coverage==1.2.0
PyYAML==6.0.1